Made the drag damage system/component server-only
Changed the system to work with events and timers instead of an update loop
