+++ Content.Server/Damage/Components/DamageOnDraggingCritComponent.cs
@@ -1,43 +1,46 @@
-using Content.Shared.Damage.Systems;
-using Robust.Shared.GameStates;
-using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
+using System.Threading;
+using Content.Server.Damage.Systems;
+using Content.Shared.Damage;
 
-namespace Content.Shared.Damage.Components;
+namespace Content.Server.Damage.Components;
 
-[RegisterComponent, NetworkedComponent]
+[RegisterComponent]
 [AutoGenerateComponentState]
 [Access(typeof(DamageOnDraggingCritSystem))]
 public sealed partial class DamageOnDraggingCritComponent : Component
 {
+    /// <summary>
+    /// Whether or not to check for dragging damage on this entity
+    /// </summary>
     [DataField("enabled"), ViewVariables(VVAccess.ReadWrite)]
-    [AutoNetworkedField]
     public bool Enabled = true;
 
     /// <summary>
     /// The amount of time between drag damage checks.
     /// </summary>
     [DataField("interval"), ViewVariables(VVAccess.ReadWrite)]
-    public float Interval = 0.25f;
+    public TimeSpan Interval = TimeSpan.FromSeconds(0.25);
 
     /// <summary>
-    /// The amount of time elapsed since the last damage check
+    /// The last time interval we checked for drag damage. Used to
+    /// scale Damage with the time between LastInterval and CurTime.
     /// </summary>
-    [DataField("intervalTimer"), ViewVariables(VVAccess.ReadOnly)]
-    public float IntervalTimer = 0.0f;
+    public TimeSpan LastInterval = default!;
+
+    public CancellationTokenSource? TimerCancel;
 
     /// <summary>
     /// The distance this mob has been dragged since the last
     /// damage check
     /// </summary>
-    [DataField("distanceDragged"), ViewVariables(VVAccess.ReadOnly)]
+    [ViewVariables(VVAccess.ReadOnly)]
     public float DistanceDragged = 0.0f;
 
     /// <summary>
     /// The amount of damage to add per second of being dragged.
-    /// This is scaled by IntervalTimer.
+    /// This is scaled by `CurTime - LastInterval`.
     /// </summary>
     [DataField("damage", required: true), ViewVariables(VVAccess.ReadWrite)]
-    [AutoNetworkedField]
     public DamageSpecifier Damage = default!;
 
     /// <summary>
@@ -47,6 +50,5 @@ public sealed partial class DamageOnDraggingCritComponent : Component
     /// This is scaled by Interval.
     /// </summary>
     [DataField("distanceThreshold"), ViewVariables(VVAccess.ReadWrite)]
-    [AutoNetworkedField]
     public float DistanceThreshold = 2.2f;
 }
